+++ beers/serializers.py
@@ -6,22 +6,24 @@ from beers.models.beers import Beer
 
 
 class BeerSerializer(serializers.HyperlinkedModelSerializer):
-    bar = serializers.CharField(source="bar.name", read_only=True)
+    bar_pk = serializers.CharField(source="bar.pk", read_only=True)
+    bar_name = serializers.CharField(source="bar.name", read_only=True)
 
     class Meta:
         model = Beer
-        fields = "__all__"
+        fields = ["pk", "name", "price", "description", "specifications", "bar_name", "bar_pk"]
 
 
 class BarSerializer(serializers.HyperlinkedModelSerializer):
     class Meta:
         model = Bar
-        exclude = ("updated",)
+        fields = ["pk", "name", "website"]
 
 
 class BarBranchSerializer(serializers.HyperlinkedModelSerializer):
-    bar = serializers.CharField(source="bar.name", read_only=True)
+    bar_pk = serializers.CharField(source="bar.pk", read_only=True)
+    bar_name = serializers.CharField(source="bar.name", read_only=True)
 
     class Meta:
         model = BarBranch
-        fields = "__all__"
+        fields = ["pk", "address", "metro", "bar_name", "bar_pk", "latitude", "longitude"]
