+++ beers/views/base.py
@@ -6,7 +6,7 @@ from beers.logics import sync_to_db
 
 @staff_required
 def sync_all_beers(request) -> HttpResponse:
-    sync_to_db.sync_lambic()
     sync_to_db.sync_we_cidreria()
     sync_to_db.sync_we_pub()
+    sync_to_db.sync_lambic()
     return HttpResponse("Все ок!")
